[Decomp][Spell] Cohesion split: extract SpellPower handlers to SpellPower.cpp
