+++ src/game/WorldHandlers/Spell.cpp
@@ -2826,254 +2826,11 @@ void Spell::prepare(SpellCastTargets const* targets, Aura* triggeredByAura)
 
 
 
-/**
- * @brief Consumes or updates the cast item after spell use when required.
- */
-void Spell::TakeCastItem()
-{
-    if (!m_CastItem || m_caster->GetTypeId() != TYPEID_PLAYER)
-    {
-        return;
-    }
-
-    // not remove cast item at triggered spell (equipping, weapon damage, etc)
-    if (m_IsTriggeredSpell && !(m_targets.m_targetMask & TARGET_FLAG_TRADE_ITEM))
-    {
-        return;
-    }
-
-    ItemPrototype const* proto = m_CastItem->GetProto();
-
-    if (!proto)
-    {
-        // This code is to avoid a crash
-        // I'm not sure, if this is really an error, but I guess every item needs a prototype
-        sLog.outError("Cast item (%s) has no item prototype", m_CastItem->GetGuidStr().c_str());
-        return;
-    }
-
-    bool expendable = false;
-    bool withoutCharges = false;
-
-    for (int i = 0; i < MAX_ITEM_PROTO_SPELLS; ++i)
-    {
-        if (proto->Spells[i].SpellId)
-        {
-            // item has limited charges
-            if (proto->Spells[i].SpellCharges)
-            {
-                if (proto->Spells[i].SpellCharges < 0 && !(proto->ExtraFlags & ITEM_EXTRA_NON_CONSUMABLE))
-                {
-                    expendable = true;
-                }
-
-                int32 charges = m_CastItem->GetSpellCharges(i);
-
-                // item has charges left
-                if (charges)
-                {
-                    (charges > 0) ? --charges : ++charges;  // abs(charges) less at 1 after use
-                    if (proto->Stackable == 1)
-                    {
-                        m_CastItem->SetSpellCharges(i, charges);
-                    }
-                    m_CastItem->SetState(ITEM_CHANGED, (Player*)m_caster);
-                }
-
-                // all charges used
-                withoutCharges = (charges == 0);
-            }
-        }
-    }
-
-    if (expendable && withoutCharges)
-    {
-        uint32 count = 1;
-        ((Player*)m_caster)->DestroyItemCount(m_CastItem, count, true);
-
-        // prevent crash at access to deleted m_targets.getItemTarget
-        ClearCastItem();
-    }
-}
-
-/**
- * @brief Deducts the spell power cost from the caster.
- */
-void Spell::TakePower()
-{
-    if (m_CastItem || m_triggeredByAuraSpell || m_IsTriggeredSpell)
-    {
-        return;
-    }
-
-    // health as power used
-    if (m_spellInfo->powerType == POWER_HEALTH)
-    {
-        m_caster->ModifyHealth(-(int32)m_powerCost);
-        return;
-    }
-
-    if (m_spellInfo->powerType >= MAX_POWERS)
-    {
-        sLog.outError("Spell::TakePower: Unknown power type '%d'", m_spellInfo->powerType);
-        return;
-    }
-
-    Powers powerType = Powers(m_spellInfo->powerType);
-
-    m_caster->ModifyPower(powerType, -(int32)m_powerCost);
-
-    // Set the five second timer
-    if (powerType == POWER_MANA && m_powerCost > 0)
-    {
-        m_caster->SetLastManaUse();
-    }
-}
 
 
 
-/**
- * @brief Consumes spell reagents from the player caster inventory.
- */
-void Spell::TakeReagents()
-{
-    if (m_caster->GetTypeId() != TYPEID_PLAYER)
-    {
-        return;
-    }
 
-    if (IgnoreItemRequirements())                           // reagents used in triggered spell removed by original spell or don't must be removed.
-    {
-        return;
-    }
 
-    Player* p_caster = (Player*)m_caster;
-    if (p_caster->CanNoReagentCast(m_spellInfo))
-    {
-        return;
-    }
-
-    for (uint32 x = 0; x < MAX_SPELL_REAGENTS; ++x)
-    {
-        if (m_spellInfo->Reagent[x] <= 0)
-        {
-            continue;
-        }
-
-        uint32 itemid = m_spellInfo->Reagent[x];
-        uint32 itemcount = m_spellInfo->ReagentCount[x];
-
-        // if CastItem is also spell reagent
-        if (m_CastItem)
-        {
-            ItemPrototype const* proto = m_CastItem->GetProto();
-            if (proto && proto->ItemId == itemid)
-            {
-                for (int s = 0; s < MAX_ITEM_PROTO_SPELLS; ++s)
-                {
-                    // CastItem will be used up and does not count as reagent
-                    int32 charges = m_CastItem->GetSpellCharges(s);
-                    if (proto->Spells[s].SpellCharges < 0 && abs(charges) < 2)
-                    {
-                        ++itemcount;
-                        break;
-                    }
-                }
-
-                m_CastItem = NULL;
-            }
-        }
-
-        // if getItemTarget is also spell reagent
-        if (m_targets.getItemTargetEntry() == itemid)
-        {
-            m_targets.setItemTarget(NULL);
-        }
-
-        p_caster->DestroyItemCount(itemid, itemcount, true);
-    }
-}
-
-/**
- * @brief Applies additional configured threat from spell_threat data.
- */
-void Spell::HandleThreatSpells()
-{
-    if (m_UniqueTargetInfo.empty())
-    {
-        return;
-    }
-
-    SpellThreatEntry const* threatEntry = sSpellMgr.GetSpellThreatEntry(m_spellInfo->Id);
-
-    if (!threatEntry || (!threatEntry->threat && threatEntry->ap_bonus == 0.0f))
-    {
-        return;
-    }
-
-    float threat = threatEntry->threat;
-    if (threatEntry->ap_bonus != 0.0f)
-    {
-        threat += threatEntry->ap_bonus * m_caster->GetTotalAttackPowerValue(GetWeaponAttackType(m_spellInfo));
-    }
-
-    bool positive = true;
-    uint8 effectMask = 0;
-    for (int i = 0; i < MAX_EFFECT_INDEX; ++i)
-    {
-        if (m_spellInfo->Effect[i])
-        {
-            effectMask |= (1 << i);
-        }
-    }
-
-    if (m_negativeEffectMask & effectMask)
-    {
-        // can only handle spells with clearly defined positive/negative effect, check at spell_threat loading probably not perfect
-        // so abort when only some effects are negative.
-        if ((m_negativeEffectMask & effectMask) != effectMask)
-        {
-            DEBUG_FILTER_LOG(LOG_FILTER_SPELL_CAST, "Spell %u, rank %u, is not clearly positive or negative, ignoring bonus threat", m_spellInfo->Id, sSpellMgr.GetSpellRank(m_spellInfo->Id));
-            return;
-        }
-        positive = false;
-    }
-
-    // since 2.0.1 threat from positive effects also is distributed among all targets, so the overall caused threat is at most the defined bonus
-    threat /= m_UniqueTargetInfo.size();
-
-    for (TargetList::const_iterator ihit = m_UniqueTargetInfo.begin(); ihit != m_UniqueTargetInfo.end(); ++ihit)
-    {
-        if (ihit->missCondition != SPELL_MISS_NONE)
-        {
-            continue;
-        }
-
-        Unit* target = m_caster->GetObjectGuid() == ihit->targetGUID ? m_caster : sObjectAccessor.GetUnit(*m_caster, ihit->targetGUID);
-        if (!target)
-        {
-            continue;
-        }
-
-        // positive spells distribute threat among all units that are in combat with target, like healing
-        if (positive)
-        {
-            target->GetHostileRefManager().threatAssist(m_caster /*real_caster ??*/, threat, m_spellInfo);
-        }
-        // for negative spells threat gets distributed among affected targets
-        else
-        {
-            if (!target->CanHaveThreatList())
-            {
-                continue;
-            }
-
-            target->AddThreat(m_caster, threat, false, GetSpellSchoolMask(m_spellInfo), m_spellInfo);
-        }
-    }
-
-    DEBUG_FILTER_LOG(LOG_FILTER_SPELL_CAST, "Spell %u added an additional %f threat for %s %zu target(s)", m_spellInfo->Id, threat, positive ? "assisting" : "harming", m_UniqueTargetInfo.size());
-}
